+++ src/vrm/gltf.rs
@@ -1,2 +1,9 @@
 pub mod extensions;
 pub mod materials;
+
+pub mod prelude {
+    pub use crate::vrm::gltf::{
+        extensions::{vrmc_spring_bone::*, vrmc_vrm::*, VrmExtensions, VrmNode},
+        materials::*,
+    };
+}