+++ src/vrm/mtoon.rs
@@ -1,17 +1,26 @@
 mod material;
-mod outline;
+mod outline_pass;
 mod setup;
 
 use crate::vrm::gltf::materials::VrmcMaterialsExtensitions;
-use crate::vrm::mtoon::outline::MToonOutlinePlugin;
+use crate::vrm::mtoon::outline_pass::MToonOutlinePlugin;
 use crate::vrm::mtoon::setup::MToonMaterialSetupPlugin;
 use bevy::asset::{load_internal_asset, weak_handle, AssetId};
 use bevy::prelude::*;
 use std::collections::HashMap;
 
-pub use material::*;
+use crate::prelude::*;
 
-const MTOON_SHADER_HANDLE: Handle<Shader> = weak_handle!("9a96eff2-1676-1dc0-9abc-2fd5e7134441");
+pub mod prelude {
+    pub use crate::vrm::mtoon::{material::prelude::*, MtoonMaterialPlugin, VrmcMaterialRegistry};
+}
+
+const MTOON_FRAGMENT_SHADER_HANDLE: Handle<Shader> =
+    weak_handle!("9a96eff2-1676-1dc0-9abc-2fd5e7134441");
+const MTOON_VERTEX_SHADER_HANDLE: Handle<Shader> =
+    weak_handle!("f4041db8-c464-b84c-e3c9-e618527945a1");
+const MTOON_TYPES_SHADER_HANDLE: Handle<Shader> =
+    weak_handle!("5d9302a3-6498-9d2a-fadb-842d01c87697");
 
 pub struct MtoonMaterialPlugin;
 
@@ -26,11 +35,28 @@ impl Plugin for MtoonMaterialPlugin {
         #[cfg(feature = "reflect")]
         {
             app.register_type::<MToonMaterial>()
-                .register_type::<crate::vrm::mtoon::outline::MToonOutline>()
+                .register_type::<MToonOutline>()
                 .register_type::<VrmcMaterialRegistry>();
         }
 
-        load_internal_asset!(app, MTOON_SHADER_HANDLE, "mtoon.wgsl", Shader::from_wgsl);
+        load_internal_asset!(
+            app,
+            MTOON_FRAGMENT_SHADER_HANDLE,
+            "mtoon_fragment.wgsl",
+            Shader::from_wgsl
+        );
+        load_internal_asset!(
+            app,
+            MTOON_VERTEX_SHADER_HANDLE,
+            "mtoon_vertex.wgsl",
+            Shader::from_wgsl
+        );
+        load_internal_asset!(
+            app,
+            MTOON_TYPES_SHADER_HANDLE,
+            "mtoon_types.wgsl",
+            Shader::from_wgsl
+        );
     }
 }