+++ src/vrm/mtoon/material.rs
@@ -1,8 +1,10 @@
+mod outline;
 mod rim_lighting;
 mod shade;
 mod uv_animation;
 
-use crate::vrm::mtoon::MTOON_SHADER_HANDLE;
+use crate::vrm::mtoon::material::outline::{MToonOutline, OutlineWidthMode};
+use crate::vrm::mtoon::{MTOON_FRAGMENT_SHADER_HANDLE, MTOON_VERTEX_SHADER_HANDLE};
 use bevy::math::Affine2;
 use bevy::pbr::{MaterialPipeline, MaterialPipelineKey, OpaqueRendererMethod};
 use bevy::prelude::*;
@@ -17,6 +19,16 @@ use bitflags::bitflags;
 pub use rim_lighting::RimLighting;
 pub use shade::Shade;
 pub use uv_animation::UVAnimation;
+
+pub mod prelude {
+    pub use crate::vrm::mtoon::material::{
+        outline::{MToonOutline, OutlineWidthMode},
+        rim_lighting::RimLighting,
+        shade::Shade,
+        uv_animation::UVAnimation,
+        MToonMaterial, MToonMaterialKey,
+    };
+}
 
 /// [VRMC_materials_mtoon-1.0](https://github.com/vrm-c/vrm-specification/blob/master/specification/VRMC_materials_mtoon-1.0/README.md)
 #[cfg_attr(feature = "reflect", derive(Reflect))]
@@ -61,10 +73,15 @@ pub struct MToonMaterial {
     #[sampler(114)]
     #[dependency]
     pub emissive_texture: Option<Handle<Image>>,
+    #[texture(115)]
+    #[sampler(116)]
+    #[dependency]
+    pub outline_width_multiply_texture: Option<Handle<Image>>,
     pub uv_animation: UVAnimation,
     pub uv_transform: Affine2,
     pub rim_lighting: RimLighting,
     pub shade: Shade,
+    pub outline: MToonOutline,
     pub base_color: Color,
     /// [VRMC_materials_mtoon-1.0](https://github.com/vrm-c/vrm-specification/blob/master/specification/VRMC_materials_mtoon-1.0/README.md#emission)
     pub emissive: LinearRgba,
@@ -75,6 +92,8 @@ pub struct MToonMaterial {
     /// [VRMC_materials_mtoon-1.0](https://github.com/vrm-c/vrm-specification/blob/master/specification/VRMC_materials_mtoon-1.0/README.md#renderqueueoffsetnumber)
     pub depth_bias: f32,
     pub opaque_renderer_method: OpaqueRendererMethod,
+    pub render_queue_offset: f32,
+    pub transparent_with_z_write: bool,
     #[cfg_attr(feature = "reflect", reflect(ignore, clone))]
     pub cull_mode: Option<Face>,
 }
@@ -85,12 +104,17 @@ bitflags! {
     pub struct MToonMaterialKey: u64 {
         const CULL_FRONT = 1 << 0;
         const CULL_BACK = 1 << 1;
+        const TRANSPARENT_WITH_Z_WRITE = 1 << 2;
     }
 }
 
 impl Material for MToonMaterial {
+    fn vertex_shader() -> ShaderRef {
+        MTOON_VERTEX_SHADER_HANDLE.into()
+    }
+
     fn fragment_shader() -> ShaderRef {
-        MTOON_SHADER_HANDLE.into()
+        MTOON_FRAGMENT_SHADER_HANDLE.into()
     }
 
     fn alpha_mode(&self) -> AlphaMode {
@@ -102,7 +126,19 @@ impl Material for MToonMaterial {
     }
 
     fn depth_bias(&self) -> f32 {
-        self.depth_bias
+        let bias = if matches!(self.alpha_mode, AlphaMode::Blend) {
+            self.depth_bias + self.render_queue_offset
+        } else {
+            self.depth_bias
+        };
+        let offset = match (self.alpha_mode, self.transparent_with_z_write) {
+            (AlphaMode::Opaque, _) => -10000.,
+            (AlphaMode::Mask(_), _) => -1000.,
+            (AlphaMode::Blend, true) => -100.,
+            (AlphaMode::Blend, false) => -10.0,
+            _ => 0.,
+        };
+        bias + offset
     }
 
     fn specialize(
@@ -119,6 +155,14 @@ impl Material for MToonMaterial {
             } else {
                 None
             };
+        if let Some(stencil) = descriptor.depth_stencil.as_mut() {
+            if key
+                .bind_group_data
+                .intersects(MToonMaterialKey::TRANSPARENT_WITH_Z_WRITE)
+            {
+                stencil.depth_write_enabled = true;
+            }
+        }
         Ok(())
     }
 }
@@ -134,6 +178,10 @@ impl From<&MToonMaterial> for MToonMaterialKey {
             MToonMaterialKey::CULL_BACK,
             material.cull_mode == Some(Face::Back),
         );
+        key.set(
+            MToonMaterialKey::TRANSPARENT_WITH_Z_WRITE,
+            matches!(material.alpha_mode, AlphaMode::Blend) && material.transparent_with_z_write,
+        );
         key
     }
 }
@@ -146,6 +194,7 @@ impl Default for MToonMaterial {
             shade_multiply_texture: None,
             rim_multiply_texture: None,
             uv_animation_mask_texture: None,
+            outline_width_multiply_texture: None,
             matcap_texture: None,
             emissive_texture: None,
             uv_animation: UVAnimation::default(),
@@ -158,8 +207,11 @@ impl Default for MToonMaterial {
             alpha_mode: AlphaMode::default(),
             double_sided: false,
             depth_bias: 0.0,
+            render_queue_offset: 0.0,
+            transparent_with_z_write: false,
             opaque_renderer_method: OpaqueRendererMethod::default(),
             cull_mode: None,
+            outline: MToonOutline::default(),
         }
     }
 }
@@ -177,10 +229,65 @@ bitflags::bitflags! {
         const DOUBLE_SIDED = 1 << 7;
         const ALPHA_MODE_MASK = 1 << 8;
         const ALPHA_MODE_ALPHA_TO_COVERAGE = 1 << 9;
+        const ALPHA_MODE_BLEND = 1 << 10;
+        const OUTLINE_WIDTH_MULTIPLY_TEXTURE = 1 << 11;
+    }
+}
+
+impl From<&MToonMaterial> for MtoonFlags {
+    fn from(value: &MToonMaterial) -> Self {
+        let mut flags = MtoonFlags::empty();
+        flags.set(
+            MtoonFlags::BASE_COLOR_TEXTURE,
+            value.base_color_texture.is_some(),
+        );
+        flags.set(MtoonFlags::DOUBLE_SIDED, value.double_sided);
+        flags.set(
+            MtoonFlags::SHADING_SHIFT_TEXTURE,
+            value.shading_shift_texture.is_some(),
+        );
+        flags.set(
+            MtoonFlags::SHADE_MULTIPLY_TEXTURE,
+            value.shade_multiply_texture.is_some(),
+        );
+        flags.set(
+            MtoonFlags::RIM_MAP_TEXTURE,
+            value.rim_multiply_texture.is_some(),
+        );
+        flags.set(
+            MtoonFlags::UV_ANIMATION_MASK_TEXTURE,
+            value.uv_animation_mask_texture.is_some(),
+        );
+        flags.set(MtoonFlags::MATCAP_TEXTURE, value.matcap_texture.is_some());
+        flags.set(
+            MtoonFlags::ALPHA_MODE_MASK,
+            matches!(value.alpha_mode, AlphaMode::Mask(_)),
+        );
+        flags.set(
+            MtoonFlags::ALPHA_MODE_ALPHA_TO_COVERAGE,
+            matches!(value.alpha_mode, AlphaMode::AlphaToCoverage),
+        );
+        flags.set(
+            MtoonFlags::ALPHA_MODE_BLEND,
+            matches!(value.alpha_mode, AlphaMode::Blend),
+        );
+        flags.set(
+            MtoonFlags::OUTLINE_WIDTH_MULTIPLY_TEXTURE,
+            value.outline_width_multiply_texture.is_some(),
+        );
+        flags
+    }
+}
+
+bitflags::bitflags! {
+    #[repr(transparent)]
+    pub struct OutlineWidthModeFlags: u32 {
+        const WORLD_COORDINATES = 1 << 0;
     }
 }
 
 #[derive(Clone, Default, ShaderType)]
+#[repr(align(16))]
 pub struct MToonMaterialUniform {
     pub flags: u32,
     pub base_color: Vec4,
@@ -201,6 +308,10 @@ pub struct MToonMaterialUniform {
     pub parametric_rim_fresnel_power: f32,
     pub rim_lighting_mix_factor: f32,
     pub alpha_cutoff: f32,
+    pub outline_flags: u32,
+    pub outline_color: Vec4,
+    pub outline_width_factor: f32,
+    pub outline_lighting_mix_factor: f32,
 }
 
 impl AsBindGroupShaderType<MToonMaterialUniform> for MToonMaterial {
@@ -208,39 +319,13 @@ impl AsBindGroupShaderType<MToonMaterialUniform> for MToonMaterial {
         &self,
         _images: &RenderAssets<GpuImage>,
     ) -> MToonMaterialUniform {
-        let mut flags = MtoonFlags::empty();
-        flags.set(
-            MtoonFlags::BASE_COLOR_TEXTURE,
-            self.base_color_texture.is_some(),
-        );
-        flags.set(MtoonFlags::DOUBLE_SIDED, self.double_sided);
-        flags.set(
-            MtoonFlags::SHADING_SHIFT_TEXTURE,
-            self.shading_shift_texture.is_some(),
-        );
-        flags.set(
-            MtoonFlags::SHADE_MULTIPLY_TEXTURE,
-            self.shade_multiply_texture.is_some(),
-        );
-        flags.set(
-            MtoonFlags::RIM_MAP_TEXTURE,
-            self.rim_multiply_texture.is_some(),
-        );
-        flags.set(
-            MtoonFlags::UV_ANIMATION_MASK_TEXTURE,
-            self.uv_animation_mask_texture.is_some(),
-        );
-        flags.set(MtoonFlags::MATCAP_TEXTURE, self.matcap_texture.is_some());
-        flags.set(
-            MtoonFlags::ALPHA_MODE_MASK,
-            matches!(self.alpha_mode, AlphaMode::Mask(_)),
-        );
-        flags.set(
-            MtoonFlags::ALPHA_MODE_ALPHA_TO_COVERAGE,
-            matches!(self.alpha_mode, AlphaMode::AlphaToCoverage),
+        let mut outline_flags = OutlineWidthModeFlags::empty();
+        outline_flags.set(
+            OutlineWidthModeFlags::WORLD_COORDINATES,
+            matches!(self.outline.mode, OutlineWidthMode::WorldCoordinates),
         );
         MToonMaterialUniform {
-            flags: flags.bits(),
+            flags: MtoonFlags::from(self).bits(),
             shade_color: self.shade.color.to_vec4(),
             shading_shift_factor: self.shade.shading_shift_factor,
             shading_shift_texture_offset: self.shade.texture_offset,
@@ -260,8 +345,12 @@ impl AsBindGroupShaderType<MToonMaterialUniform> for MToonMaterial {
             emissive_color: self.emissive.to_vec4(),
             alpha_cutoff: match self.alpha_mode {
                 AlphaMode::Mask(value) => value,
-                _ => 0.5,
+                _ => 1e-3,
             },
+            outline_flags: outline_flags.bits(),
+            outline_color: self.outline.color.to_vec4(),
+            outline_width_factor: self.outline.width_factor,
+            outline_lighting_mix_factor: self.outline.lighting_mix_factor,
         }
     }
 }