+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## Unreleased
+
+### Breaking Changes
+
+- `MToonOutline` is no longer a component; it has become part of the `MToonMaterial` fields.
+- `OutlineWidthMode` has been added as part of the field of `MToonOutline`.
+    - Currently only supports `OutlineWidthMode::WorldCoordinates` and `OutlineWidthMode::None`, and if `screenCoordinates` is passed, the outline will not be rendered.
+- Fixed the rendering order of the outline to match the spec.
+  - refer to the [here](https://github.com/vrm-c/vrm-specification/blob/master/specification/VRMC_materials_mtoon-1.0/README.md#rendering) for more details.
+
+### Bug Fixes
+
+- Fixed outline rendering 
+
 ## v0.1.2
 
 [Relsease Notes](https://github.com/not-elm/bevy_vrm1/releases/tag/v0.1.2)